feat: split ordered and unordered list renderers (#9)
* Split ordered and unordered list renderers

Since Svelte style blocks don't cascade, we can't create a renderer for
lists that will style list items. We need to instead style the list item
directly. For more complex styling like changing the list marker, it can be
helpful to know if this is an ordered list or unordered. This change differentiates
between the two and provides a renderer for each case.

- Add ordered attribute to the parser
- Use the ordered attribute to determine if the list is ordered or not
- For ordered lists, call the ordered list item renderer
- For unordered lists, call the unordered list item renderer
- Duplicates the default renderer - assigns one copy to each name (not needed but
  could help avoid confusion)
- Map the default renderers

* Make original list item the default list item

- When ordered and unordered variations are not defined, use the ListItem
- If one of those is defined, use the specialized version
- Remove the default ordered and unordered renderers, those are redundant now

see: https://github.com/pablo-abc/svelte-markdown/pull/9#discussion_r626163504

* Readme includes optional list items

- Add section for optional list items
- Clarify the fallback behavior of `listitem`
- Add a small example

* Fix bug in unordered lists

- Unordered lists were skipped accidently
- Lets not do that